Scope parsing WIP + variable parsing fixes

+++ src/arraylist.c
@@ -33,7 +33,7 @@ void arraylist_add_all(arraylist_t *list, arraylist_t *values) {
 		arraylist_add(list, arraylist_get(values, i));
 }
 
-void *arraylist_get(arraylist_t *list, unsigned int index) {
+void *arraylist_get(arraylist_t *list, size_t index) {
 	return index >= list->size ? NULL : list->array[index];
 }
